# spec/openapi.yaml
openapi: 3.1.0
info:
  title: I3X - Industrial Information Interface eXchange API
  description: |
    A common API for Contextualized Manufacturing Information Platforms (CMIPs) as defined in RFC 001.

    This specification enables portable application development across different manufacturing
    information platforms by providing a vendor-agnostic API specification.

    ## Key Concepts

    - **ElementId**: Platform-specific unique string identifier for any element
    - **Namespace**: Logical grouping of related types and instances
    - **ObjectType**: Schema definition for a class of objects
    - **ObjectInstance**: Actual equipment, sensors, or processes with attribute values
    - **VQT**: Value-Quality-Timestamp format for all data values
    - **Relationships**: Organizational (HasParent/HasChildren) and Composition (HasComponent/ComponentOf)

    ## maxDepth Parameter

    Controls recursion through HasComponent relationships:
    - `maxDepth=0`: Infinite recursion (include all nested composed elements)
    - `maxDepth=1`: No recursion (just this element's value) - default
    - `maxDepth=N`: Recurse up to N levels deep

    ## Quality Values

    - `Good`: Value is valid and current
    - `GoodNoData`: No data available but connection is good
    - `Bad`: Value is invalid or connection failed
    - `Uncertain`: Value quality cannot be determined

  version: 1.0.0
  contact:
    name: CESMII - The Smart Manufacturing Institute
    url: https://www.cesmii.org
...
servers:
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      summary: List all subscriptions
      description: |
        Returns a list of all active subscriptions for the authenticated client.

        Note: This endpoint is informational and not explicitly defined in RFC 001.
      tags:
        - Subscriptions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createSubscription
      summary: Create a subscription
      description: |
        Registers a client for a new subscription. This initial handshake allows
        the CMIP to allocate resources for a client.

        The response MUST include a Subscription ID that may be used for follow-up calls.

        Monitoring begins when objects are registered via the /register endpoint.

        **RFC Reference:** 4.2.3.1
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '200':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSubscriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /subscriptions/{subscriptionId}:
    get:
      operationId: getSubscription
      summary: Get subscription details
      description: |
        Returns details about a specific subscription including monitored items
        and current status.
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteSubscription
      summary: Delete a subscription
      description: |
        Cancels publication of future messages for the specified subscription,
        allowing the CMIP to release resources and state held for the client.

        **RFC Reference:** 4.2.3.5
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
      responses:
        '200':
          description: Subscription deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsubscribeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /subscriptions/{subscriptionId}/register:
    post:
      operationId: registerMonitoredItems
      summary: Register monitored items
      description: |
        Registers the ElementIds the client wishes to subscribe to.

        Upon registration, the CMIP MUST begin publishing changed values.
        This method is additive - the client can add additional monitored items later.

        The maxDepth parameter controls recursion through HasComponent relationships
        when publishing updates for composition elements.

        **RFC Reference:** 4.2.3.2
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterItemsRequest'
      responses:
        '200':
          description: Items registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterItemsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Subscription or element not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /subscriptions/{subscriptionId}/unregister:
    post:
      operationId: unregisterMonitoredItems
      summary: Remove monitored items
      description: |
        Removes the ElementIds the client no longer wishes to monitor
        from the specified subscription.

        **RFC Reference:** 4.2.3.3
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterItemsRequest'
      responses:
        '200':
          description: Items unregistered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnregisterItemsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /subscriptions/{subscriptionId}/stream:
    get:
      operationId: streamSubscription
      summary: Open SSE stream (streaming)
      description: |
        Opens a Server-Sent Events (SSE) stream for real-time data delivery.

        This implements streaming (At Most Once) delivery - the CMIP will publish
        messages as data becomes available, but provides no guarantee of delivery.

        When the stream is active, updates are delivered immediately rather than
        being queued for /sync retrieval.

        **RFC Reference:** 4.2.3.2 (streaming)
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
        - name: mode
          in: query
          description: Delivery mode (streaming for SSE)
          required: false
          schema:
            type: string
            enum: [streaming]
            default: streaming
      responses:
        '200':
          description: SSE stream opened successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionUpdate'
              description: |
                Stream of JSON arrays, each containing subscription updates.
                Each line is a complete JSON array that can be parsed independently.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /subscriptions/{subscriptionId}/sync:
    post:
      operationId: syncSubscription
      summary: Poll for updates (sync)
      description: |
        Retrieves queued updates for the subscription and acknowledges receipt.

        This implements sync (Exactly Once) delivery - the server will persist
        messages for re-delivery until the client acknowledges successful receipt.

        The CMIP MUST respond with an array of updates for elements that have
        changed since the last Sync call. If no elements have changed, the
        response MUST be an empty array.

        **RFC Reference:** 4.2.3.4
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
        - name: mode
          in: query
          description: Delivery mode (sync for polling)
          required: false
          schema:
            type: string
            enum: [sync]
            default: sync
      responses:
        '200':
          description: Successful response with queued updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionUpdate'
              example:
                - elementId: "sensor-001"
                  value: 72.5
                  quality: "Good"
                  timestamp: "2025-01-08T10:30:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

# ==============================================================================
# Components
# ==============================================================================
components:
...
  schemas:
...
    CreateSubscriptionRequest:
      type: object
      description: Request body for creating a subscription
      properties: {}
      additionalProperties: true
    CreateSubscriptionResponse:
      type: object
      description: Response from creating a subscription
      required:
        - subscriptionId
        - message
      properties:
        subscriptionId:
          type: string
          description: Unique identifier for the new subscription
          example: "0"
        message:
          type: string
          description: Human-readable confirmation message
          example: "Subscription created successfully."
    SubscriptionList:
      type: object
      description: List of subscriptions
      required:
        - subscriptionIds
      properties:
        subscriptionIds:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionSummary'
    SubscriptionSummary:
      type: object
      description: Brief subscription information
      required:
        - subscriptionId
        - created
      properties:
        subscriptionId:
          type: string
          description: Unique subscription identifier
        created:
          type: string
          format: date-time
          description: Timestamp when subscription was created
    SubscriptionDetails:
      type: object
      description: Full subscription details
      required:
        - subscriptionId
        - created
        - isStreaming
        - queuedUpdates
        - objects
      properties:
        subscriptionId:
          type: string
        created:
          type: string
          format: date-time
        isStreaming:
          type: boolean
          description: True if SSE stream is currently active
        queuedUpdates:
          type: integer
          description: Number of updates waiting in queue
        objects:
          type: array
          items:
            type: string
          description: List of monitored element IDs
    RegisterItemsRequest:
      type: object
      description: Request body for registering or unregistering monitored items
      required:
        - elementIds
      properties:
        elementIds:
          type: array
          items:
            type: string
          description: Array of ElementIds to monitor
          example: ["sensor-001", "sensor-002"]
        maxDepth:
          type: integer
          minimum: 0
          default: 1
          description: |
            Controls recursion through HasComponent relationships:
            - 0: Infinite recursion
            - 1: No recursion (default)
            - N: Recurse up to N levels
    RegisterItemsResponse:
      type: object
      description: Response from registering monitored items
      required:
        - message
        - totalObjects
      properties:
        message:
          type: string
          example: "Registered 5 objects to subscription."
        totalObjects:
          type: integer
          description: Total number of monitored objects after registration
    UnregisterItemsResponse:
      type: object
      description: Response from unregistering monitored items
      required:
        - message
      properties:
        message:
          type: string
          example: "Unregistered 3 objects from subscription."
    UnsubscribeResponse:
      type: object
      description: Response from deleting a subscription
      required:
        - message
        - unsubscribed
        - not_found
      properties:
        message:
          type: string
          example: "Unsubscribe processed."
        unsubscribed:
          type: array
          items:
            type: string
          description: List of successfully unsubscribed IDs
        not_found:
          type: array
          items:
            type: string
          description: List of IDs that were not found
    SubscriptionUpdate:
      type: object
      description: A single update from a subscription
      required:
        - elementId
        - value
      properties:
        elementId:
          type: string
          description: ElementId of the changed element
        value:
          description: The new value
        quality:
          type: string
          description: Quality indicator
          enum:
            - Good
            - GoodNoData
            - Bad
            - Uncertain
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the change

    # --------------------------------------------------------------------------
    # Error Response
    # --------------------------------------------------------------------------
    Error:
      type: object
      description: Standard error response
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Element 'unknown-id' not found"

  # ============================================================================
  # Parameters
  # ============================================================================
  parameters:
...
    SubscriptionIdPath:
      name: subscriptionId
      in: path
      required: true
      description: Unique subscription identifier
      schema:
        type: string
      example: "0"
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid request body"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Authentication required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Element 'unknown-id' not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Internal server error"

  # ============================================================================
  # Security Schemes
  # ============================================================================
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (RFC 5.3.1)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT/OAuth bearer token authentication (optional)
security:
  - ApiKeyAuth: []
  - BearerAuth: []